+++ extern/OptiML/compiler/src/ops/SumOpsExp.scala
@@ -19,8 +19,8 @@ import optiml.compiler.ops._
 trait SumOpsExp extends SumOps with BaseFatExp with EffectExp {
   this: OptiMLExp =>
 
-   case class Sum[A:Manifest:Arith](start: Exp[Int], end: Exp[Int], map: Exp[Int] => Exp[A], init: Exp[A])(implicit canSum: CanSum[A,A], ctx: SourceContext)
-     extends DeliteOpMapReduce[Int,A] {
+   case class Sum[A:Manifest:Arith](start: Exp[Long], end: Exp[Long], map: Exp[Long] => Exp[A], init: Exp[A])(implicit canSum: CanSum[A,A], ctx: SourceContext)
+     extends DeliteOpMapReduce[Long,A] {
 
      override val mutable = true // can we do this automatically?
 
@@ -36,7 +36,7 @@ trait SumOpsExp extends SumOps with BaseFatExp with EffectExp {
      def sc = implicitly[SourceContext]
    }
 
-  case class SumIf[R:Manifest:Arith,A:Manifest](start: Exp[Int], end: Exp[Int], co: Exp[Int] => Exp[Boolean], fu: Exp[Int] => Exp[A])(implicit canSum: CanSum[R,A], ctx: SourceContext) // TODO aks: CS into Arith
+  case class SumIf[R:Manifest:Arith,A:Manifest](start: Exp[Long], end: Exp[Long], co: Exp[Long] => Exp[Boolean], fu: Exp[Long] => Exp[A])(implicit canSum: CanSum[R,A], ctx: SourceContext) // TODO aks: CS into Arith
     extends DeliteOpFilterReduceFold[R] {
 
     override val mutable = true // can we do this automatically?
@@ -61,12 +61,12 @@ trait SumOpsExp extends SumOps with BaseFatExp with EffectExp {
 
 
     // this is the reduce used inside each chunk (R,A) => R
-    def reduceSeq = (a,b) => (reifyEffects(if (co(b._2)) { if (a._2 >= unit(0)) canSum.accA(a._1, fu(b._2)) else canSum.mutableA(fu(b._2)) } else a._1),
+    def reduceSeq = (a,b) => (reifyEffects(if (co(b._2)) { if (a._2 >= unit(0l)) canSum.accA(a._1, fu(b._2)) else canSum.mutableA(fu(b._2)) } else a._1),
                               reifyEffects(if (co(b._2)) b._2 else a._2 )) // would not work in parallel...  // returns the current index (v) if the condition is true, or a._2, which is defaulted to -1 (uninitialized)
 
     // this is the reduce used in the tree (R,R) => R
-    def reducePar = (a,b) => (reifyEffects(if (b._2 >= unit(0)) { if (a._2 >= unit(0)) canSum.accR(a._1, b._1) else canSum.mutableR(b._1) } else a._1),
-                              reifyEffects(if (b._2 >= unit(0)) b._2 else a._2))
+    def reducePar = (a,b) => (reifyEffects(if (b._2 >= unit(0l)) { if (a._2 >= unit(0l)) canSum.accR(a._1, b._1) else canSum.mutableR(b._1) } else a._1),
+                              reifyEffects(if (b._2 >= unit(0l)) b._2 else a._2))
 
     def mR = manifest[R]
     def a = implicitly[Arith[R]]
@@ -75,13 +75,13 @@ trait SumOpsExp extends SumOps with BaseFatExp with EffectExp {
     def sc = implicitly[SourceContext]
   }
 
-  def optiml_sum[A:Manifest:Arith](start: Exp[Int], end: Exp[Int], block: Exp[Int] => Exp[A])(implicit cs: CanSum[A,A], ctx: SourceContext) = {
+  def optiml_sum[A:Manifest:Arith](start: Exp[Long], end: Exp[Long], block: Exp[Long] => Exp[A])(implicit cs: CanSum[A,A], ctx: SourceContext) = {
     // don't add it back in, just re-compute it to avoid the peeled iteration problems
     val zero = block(start)
     reflectPure(Sum(start,end,block,zero))
   }
 
-  def optiml_sumif[R:Manifest:Arith,A:Manifest](start: Exp[Int], end: Exp[Int], cond: Exp[Int] => Exp[Boolean], block: Exp[Int] => Exp[A])(implicit cs: CanSum[R,A], ctx: SourceContext) = {
+  def optiml_sumif[R:Manifest:Arith,A:Manifest](start: Exp[Long], end: Exp[Long], cond: Exp[Long] => Exp[Boolean], block: Exp[Long] => Exp[A])(implicit cs: CanSum[R,A], ctx: SourceContext) = {
     reflectPure(SumIf[R,A](start, end, cond, block))
   }
